requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Forex Trading & Market Data
oandapyV20>=0.6.3
ccxt>=4.0.0

# Technical Analysis
TA-Lib>=0.4.25

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# AI Services
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.8.0

# Telegram Bot
python-telegram-bot>=20.0

# Database
sqlite3

# Async Programming
asyncio

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3

# Logging
colorlog>=6.7.0

# Data Validation
pydantic>=2.0.0

# Configuration Management
pyyaml>=6.0

# Mathematical Operations
scipy>=1.11.0

# Date/Time Handling
python-dateutil>=2.8.2

# HTTP Client
aiohttp>=3.8.0

# For backward compatibility
enum34; python_version<"3.4"
